WIP - Factor out timer code into a timer module
